requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
pymongo==3.12.1
redis==3.5.3
python-dotenv==0.19.2
bcrypt==3.2.0
gunicorn==20.1.0